Display all storage pools for migration
findStoragePoolsForMigration limits the pool to only those it thinks are 'suitable'. In reality, when the VM is powered down, you can migrate to any storage pool. In CloudMonkey this is already possible, so why not have the UI do it too? This PR fixes that. It allows for easy switching between different offerings / pools.